+++ web/app/areas/page.tsx
@@ -1,0 +1,5 @@
+import AreasModule from '../../components/areas/AreasModule';
+
+export default function AreasPage() {
+  return <AreasModule />;
+}